requirements.txt
torch
pillow
requests
langchain-core
transformers
qwen-vl-utils